--- x-pack/plugins/actions/server/builtin_action_types/connectors/jira/config.ts
@@ -1,13 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License;
- * you may not use this file except in compliance with the Elastic License.
- */
-
-import { ConnectorConfiguration } from '../types';
-import * as i18n from './translations';
-
-export const config: ConnectorConfiguration = {
-  id: '.jira',
-  name: i18n.NAME,
-};
