+++ x-pack/plugins/actions/server/builtin_action_types/jira/index.ts
@@ -4,7 +4,7 @@
  * you may not use this file except in compliance with the Elastic License.
  */
 
-import { createConnector } from '../utils';
+import { createConnector } from '../common/utils';
 
 import { api } from './api';
 import { config } from './config';
@@ -12,7 +12,7 @@ import { validate } from './validators';
 import { createExternalService } from './service';
 import { JiraSecretConfiguration, JiraPublicConfiguration } from './schema';
 
-export const connector = createConnector({
+export const getActionType = createConnector({
   api,
   config,
   validate,
