+++ x-pack/plugins/actions/server/builtin_action_types/jira/translations.ts
