+++ x-pack/plugins/actions/server/builtin_action_types/servicenow/types.ts
@@ -4,4 +4,7 @@
  * you may not use this file except in compliance with the Elastic License.
  */
 
-export { api } from '../api';
+export {
+  ConnectorPublicConfigurationType as ServiceNowPublicConfigurationType,
+  ConnectorSecretConfigurationType as ServiceNowSecretConfigurationType,
+} from '../common/types';
