+++ x-pack/plugins/siem/public/pages/case/components/configure_cases/index.test.tsx
@@ -186,6 +186,16 @@ describe('ConfigureCases', () => {
           id: '.servicenow',
           name: 'ServiceNow',
           enabled: true,
+          logo: 'test-file-stub',
+          enabledInConfig: true,
+          enabledInLicense: true,
+          minimumLicenseRequired: 'platinum',
+        },
+        {
+          id: '.jira',
+          name: 'Jira',
+          logo: 'test-file-stub',
+          enabled: true,
           enabledInConfig: true,
           enabledInLicense: true,
           minimumLicenseRequired: 'platinum',
